refactor(dock): remove Dock component and inline DockItemData type
